2/25/22 Changes: Java created and moved the files Adjunct, Admin, Faculty, and User into a folder called users. This folder is on the same level as the src folder and is now considered a package. (dont know why) Java also moved Controller from the src folder into a sample folder on the same level as src. This also made this new Controller folder a package. (also dont know why). These changes removed all the errors Java was sensing (even though the program ran "correctly" with these unresolved issues). I also added comments where Code analysis determined there were warnings.
