--- classyScheduleJavafx/src/sample/Controller.java
@@ -1,4 +1,0 @@
-package sample;
-
-public class Controller {
-}
